+++ packages/odata-service-writer/src/updates.ts
@@ -1,4 +1,3 @@
-import { render } from 'ejs';
 import type { Editor } from 'mem-fs-editor';
 import { join, normalize, posix, sep } from 'path';
 import { t } from './i18n';
@@ -9,12 +8,199 @@ import type { Manifest, ManifestNamespace } from '@sap-ux/project-access';
 import { DirName, getMinimumUI5Version, getWebappPath, hasUI5CliV3 } from '@sap-ux/project-access';
 
 /**
+ * Enhances manifest.json models with given service data.
+ *
+ * @param {string} serviceName - name of the OData service instance
+ * @param {string} serviceVersion - version of the OData service instance
+ * @param {string} serviceModel - model of the OData service instance
+ * @param {boolean} includeSynchronizationMode - whether to include synchronizationMode for model settings
+ * @param {Manifest} manifest - the manifest.json of the application
+ */
+function enhanceManifestModels(
+    serviceName: string,
+    serviceVersion: string,
+    serviceModel: string,
+    includeSynchronizationMode: boolean,
+    manifest: Manifest
+): void {
+    const models = manifest?.['sap.ui5']?.models ?? {};
+    let modelSettings: ManifestNamespace.Ui5Setting = {};
+    if (serviceVersion === '4') {
+        if (includeSynchronizationMode) {
+            modelSettings['synchronizationMode'] = 'None';
+        }
+        modelSettings['operationMode'] = 'Server';
+        modelSettings['autoExpandSelect'] = true;
+        modelSettings['earlyRequests'] = true;
+    }
+    if (models[serviceModel]?.settings) {
+        // Merge settings for existing model
+        modelSettings = {
+            ...models[serviceModel].settings,
+            ...modelSettings
+        };
+        models[serviceModel] = {
+            ...models[serviceModel],
+            dataSource: serviceName,
+            preload: true,
+            settings: modelSettings
+        };
+    } else {
+        models[serviceModel] = {
+            dataSource: serviceName,
+            preload: true,
+            settings: modelSettings
+        };
+    }
+    if (manifest['sap.ui5']) {
+        manifest['sap.ui5'] = {
+            ...manifest['sap.ui5'],
+            models
+        };
+    } else {
+        Object.assign(manifest, { 'sap.ui5': { models: { ...models } } });
+    }
+}
+
+/**
+ * Enhances manifest.json dataSources with given service data.
+ *
+ * @param {string} serviceName - name of the OData service instance
+ * @param {string} servicePath - path of the OData service instance
+ * @param {string} serviceVersion - version of the OData service instance
+ * @param {Manifest} manifest - the manifest.json of the application
+ * @param {string} serviceMetadata - metdata of the OData service instance
+ * @param {EdmxAnnotationsInfo | EdmxAnnotationsInfo[]} serviceRemoteAnnotations - remote annotations of the OData service instance
+ * @param {string | string[]} serviceLocalAnnotations - local annotations of the OData service instance
+ */
+function enhanceManifestDatasources(
+    serviceName: string,
+    servicePath: string,
+    serviceVersion: string,
+    manifest: Manifest,
+    serviceMetadata?: string,
+    serviceRemoteAnnotations?: EdmxAnnotationsInfo | EdmxAnnotationsInfo[],
+    serviceLocalAnnotations?: string | string[]
+): void {
+    const dataSources = manifest?.['sap.app'].dataSources ?? {};
+    // Service annotation names to be stored in service settings of dataSource
+    const annotations: string[] = [];
+    // Annotation dataSources used by service - remote and local annotations are handled differently
+    const annotationDataSources: { [k: string]: ManifestNamespace.DataSource } = {};
+    // Handle remote annotations used by service
+    if (Array.isArray(serviceRemoteAnnotations)) {
+        serviceRemoteAnnotations.forEach((remoteAnnotation) => {
+            if (remoteAnnotation.name) {
+                annotations.push(remoteAnnotation.name);
+                annotationDataSources[remoteAnnotation.name] = {
+                    uri: `/sap/opu/odata/IWFND/CATALOGSERVICE;v=2/Annotations(TechnicalName='${encodeURIComponent(
+                        remoteAnnotation.technicalName
+                    )}',Version='0001')/$value/`,
+                    type: 'ODataAnnotation',
+                    settings: {
+                        localUri: `localService/${serviceName}/${remoteAnnotation.technicalName}.xml`
+                    }
+                };
+            }
+        });
+    } else if (serviceRemoteAnnotations?.name) {
+        annotations.push(serviceRemoteAnnotations.name);
+        annotationDataSources[serviceRemoteAnnotations.name] = {
+            uri: `/sap/opu/odata/IWFND/CATALOGSERVICE;v=2/Annotations(TechnicalName='${encodeURIComponent(
+                serviceRemoteAnnotations.technicalName
+            )}',Version='0001')/$value/`,
+            type: 'ODataAnnotation',
+            settings: {
+                localUri: `localService/${serviceName}/${serviceRemoteAnnotations.technicalName}.xml`
+            }
+        };
+    }
+    // Handle local annotations used by service
+    if (Array.isArray(serviceLocalAnnotations)) {
+        serviceLocalAnnotations.forEach((localAnnotation: string) => {
+            annotations.push(localAnnotation);
+            annotationDataSources[localAnnotation] = {
+                type: 'ODataAnnotation',
+                uri: `annotations/${localAnnotation}.xml`,
+                settings: {
+                    localUri: `annotations/${localAnnotation}.xml`
+                }
+            };
+        });
+    } else if (serviceLocalAnnotations) {
+        annotations.push(serviceLocalAnnotations);
+        annotationDataSources[serviceLocalAnnotations] = {
+            type: 'ODataAnnotation',
+            uri: `annotations/${serviceLocalAnnotations}.xml`,
+            settings: {
+                localUri: `annotations/${serviceLocalAnnotations}.xml`
+            }
+        };
+    }
+    const settings = {
+        annotations
+    };
+    if (serviceMetadata) {
+        settings['localUri'] = `localService/${serviceName}/metadata.xml`;
+    }
+    if (serviceVersion === '4') {
+        settings['odataVersion'] = '4.0';
+    } else if (serviceVersion === '2') {
+        settings['odataVersion'] = '2.0';
+    }
+    // Create or update service dataSource in manifest.json for service
+    dataSources[serviceName] = {
+        uri: servicePath,
+        type: 'OData',
+        settings
+    };
+    // Create or update service annotation dataSources in manifest.json for service
+    for (const name in annotationDataSources) {
+        const annotationDataSource = annotationDataSources[name];
+        dataSources[name] = annotationDataSource;
+    }
+    // Update manifest.json dataSources
+    manifest['sap.app'].dataSources = dataSources;
+}
+
+/**
+ * Enhances (creates or updates existing) manifest.json with service data.
+ *
+ * @param {OdataService} service - the OData service instance
+ * @param {Manifest} manifest - the manifest.json of the application
+ */
+function enhanceManifest(service: OdataService, manifest: Manifest): void {
+    const minimumUi5Version = getMinimumUI5Version(manifest);
+    // Enhance model settings for service
+    const serviceSettings = Object.assign(service, getModelSettings(minimumUi5Version));
+    if (serviceSettings.name && serviceSettings.path && serviceSettings.model !== undefined) {
+        enhanceManifestDatasources(
+            serviceSettings.name,
+            serviceSettings.path,
+            serviceSettings.version,
+            manifest,
+            serviceSettings.metadata,
+            serviceSettings.annotations as EdmxAnnotationsInfo | EdmxAnnotationsInfo[],
+            serviceSettings.localAnnotationsName
+        );
+        // Add or update existing service model settings for manifest.json
+        enhanceManifestModels(
+            serviceSettings.name,
+            serviceSettings.version,
+            serviceSettings.model,
+            serviceSettings.includeSynchronizationMode,
+            manifest
+        );
+    }
+}
+
+/**
  * Modifies service in manifest.json and service files in a way that is supported by multiple services.
  * If service files are defined in 'localService' folder then those files are moved to respective service folder and service configuration URI are modified in manifest.json.
  *
  * @param {string} webappPath - the webapp path of an existing UI5 application
  * @param {string} dataSourceKey - dataSource key in manifest.json
- * @param {Manifest} dataSource - dataSource configuration from manifest.json
+ * @param {ManifestNamespace.DataSource} dataSource - dataSource configuration from manifest.json
  * @param {Editor} fs - the memfs editor instance
  */
 function updateExistingService(
@@ -73,14 +259,8 @@ async function updateExistingServices(webappPath: string, manifest: Manifest, fs
  * @param basePath - the root path of an existing UI5 application
  * @param service - the OData service instance
  * @param fs - the memfs editor instance
- * @param templateRoot - root folder contain the ejs templates
  */
-export async function updateManifest(
-    basePath: string,
-    service: OdataService,
-    fs: Editor,
-    templateRoot: string
-): Promise<void> {
+export async function updateManifest(basePath: string, service: OdataService, fs: Editor): Promise<void> {
     const webappPath = await getWebappPath(basePath, fs);
     const manifestPath = join(webappPath, 'manifest.json');
     // Get component app id
@@ -96,12 +276,9 @@ export async function updateManifest(
             t('error.requiredProjectPropertyNotFound', { property: `'${appProp}'.id`, path: manifestPath })
         );
     }
-
-    const manifestJsonExt = fs.read(join(templateRoot, 'extend', `manifest.json`));
-    const manifestSettings = Object.assign(service, getModelSettings(getMinimumUI5Version(modifiedManifest)));
-    // If the service object includes ejs options, for example 'client' (see: https://ejs.co/#docs),
-    // resulting in unexpected behaviour and problems when webpacking. Passing an empty options object prevents this.
-    fs.extendJSON(manifestPath, JSON.parse(render(manifestJsonExt, manifestSettings, {})));
+    // Add or update manifest.json with service
+    enhanceManifest(service, modifiedManifest);
+    fs.writeJSON(manifestPath, modifiedManifest);
 }
 
 /**